first test with sweetAlert

+++ server/public/scripts/client.js
@@ -66,7 +66,7 @@ function saveKoala(newKoala) {
     $.ajax({
       method: 'POST',
       url: '/koalas',
-      data: saveKoala
+      data: savedKoala
     }).then(function(response) {
       console.log(response);
       getKoalas()
@@ -99,19 +99,34 @@ function deleteKoala() {
   //dom traversal to delete the actaual thing we want 
   const koalaId = $(this).parent().parent().data('id')
 
-  //sending a delete request to the server
-  $.ajax({
-    method: 'DELETE',
-    url: `/koalas/${koalaId}`
-  })
-    .then((response) => {
-      console.log('dleted a koala');
-      //getting up to date data for the koalas
-      getKoalas()
-    }).catch((error) => {
-      console.log('Error in DELETE request: ', error);
-      alert('Error with deleting a song');
+  console.log('koalaId is', koalaId);
+  
+
+
+  swal({
+    title: 'Are you sure?',
+    text: "click delete to leave this koala in the wild",
+    icon: "warning",
+    buttons: true,
+    dangerMode: true
+
+  }).then((willDelete) => {
+     //sending a delete request to the server
+    $.ajax({
+      method: 'DELETE',
+      url: `/koalas/${koalaId}`
     })
+      .then((response) => {
+        console.log('dleted a koala');
+        //getting up to date data for the koalas
+        getKoalas()
+      }).catch((error) => {
+        console.log('Error in DELETE request: ', error);
+        alert('Error with deleting a song');
+      })
+
+  });
+ 
 
 }
